+++ webpack/base.config.js
@@ -2,6 +2,7 @@
 process.traceDeprecation = true;
 const path = require("path");
 const webpack = require("webpack");
+const webpack_helpers = require("patternslib/webpack/webpack-helpers");
 
 // plugins
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
@@ -77,7 +78,15 @@ module.exports = (env) => {
                 },
                 {
                     test: /\.css$/,
-                    use: [{ loader: "style-loader" }, { loader: "css-loader" }],
+                    use: [
+                        {
+                            loader: "style-loader",
+                            options: {
+                                insert: webpack_helpers.top_head_insert,
+                            },
+                        },
+                        "css-loader",
+                    ],
                 },
             ],
         },